mvi generate thumbnails in two phases
First phase: generate a smaller version of the image
in the chosen thumbnail folder, using the selected
compression format (i.e. webp). These thumbnails are
permanently saved.

Second phase: from the previously generated thumbnail,
convert it to bgra format with premultiplied alpha,
as specified in mpv's documentation. This image is
saved in a temporary folder (usually /tmp) and it is
erased when closing the program.

This way we can save disk space by storing the images
in a compressed format. The second phase should be
fast enough because the pre-processed thumbnail is
small.